#2281 Rename package since there could be conditions for date, datetime and time (maybe if future)
