+++ src/main/java/com/codeborne/selenide/conditions/datetime/DateFormatConditionOptions.java
@@ -1,4 +1,4 @@
-package com.codeborne.selenide.conditions.date;
+package com.codeborne.selenide.conditions.datetime;
 
 import com.codeborne.selenide.Condition;
