+++ src/test/java/com/codeborne/selenide/conditions/datetime/LocalDateTimeConditionTest.java
@@ -1,11 +1,11 @@
-package com.codeborne.selenide.conditions.date;
+package com.codeborne.selenide.conditions.datetime;
 
 import com.codeborne.selenide.CheckResult;
 import com.codeborne.selenide.Driver;
 import org.junit.jupiter.api.Test;
 import org.openqa.selenium.WebElement;
 
-import java.time.LocalDate;
+import java.time.LocalDateTime;
 
 import static com.codeborne.selenide.CheckResult.Verdict.ACCEPT;
 import static com.codeborne.selenide.CheckResult.Verdict.REJECT;
@@ -15,15 +15,15 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-final class LocalDateConditionTest {
+final class LocalDateTimeConditionTest {
   private final Driver driver = mock();
   private final WebElement element = mock();
-  private final LocalDateFormatCondition formatCondition = mock();
-  private final LocalDateEqualCondition condition = new LocalDateEqualCondition(LocalDate.of(2022, 10, 11), formatCondition);
-
+  private final LocalDateTimeFormatCondition formatCondition = mock();
+  private final LocalDateTime expectedDateTime = LocalDateTime.of(2022, 10, 11, 12, 13, 14);
+  private final LocalDateTimeEqualCondition condition = new LocalDateTimeEqualCondition(expectedDateTime, formatCondition);
   @Test
-  void correctDateValueWithCorrectFormat() {
-    when(formatCondition.check(any(), any())).thenReturn(new CheckResult(ACCEPT, LocalDate.of(2022, 10, 11)));
+  void correctDateTimeValueWithCorrectFormat() {
+    when(formatCondition.check(any(), any())).thenReturn(new CheckResult(ACCEPT, expectedDateTime));
     when(formatCondition.format(any())).thenReturn("formatted date");
 
     CheckResult check = condition.check(driver, element);
@@ -32,7 +32,7 @@ final class LocalDateConditionTest {
     assertThat(check.actualValue()).isEqualTo("formatted date");
 
     verify(formatCondition).check(driver, element);
-    verify(formatCondition).format(LocalDate.of(2022, 10, 11));
+    verify(formatCondition).format(expectedDateTime);
   }
 
   @Test
@@ -46,8 +46,8 @@ final class LocalDateConditionTest {
   }
 
   @Test
-  void incorrectDateValue() {
-    when(formatCondition.check(any(), any())).thenReturn(new CheckResult(ACCEPT, LocalDate.of(2023, 10, 11)));
+  void incorrectDateTimeValue() {
+    when(formatCondition.check(any(), any())).thenReturn(new CheckResult(ACCEPT, expectedDateTime.minusDays(1)));
     when(formatCondition.format(any())).thenReturn("formatted date");
 
     CheckResult check = condition.check(driver, element);
@@ -56,13 +56,14 @@ final class LocalDateConditionTest {
     assertThat(check.actualValue()).isEqualTo("formatted date");
 
     verify(formatCondition).check(driver, element);
-    verify(formatCondition).format(LocalDate.of(2023, 10, 11));
+    verify(formatCondition).format(expectedDateTime.minusDays(1));
   }
 
   @Test
   void stringRepresentationOfCondition() {
-    var condition = new LocalDateEqualCondition(LocalDate.of(2022, 10, 11), "yyyy/MM/dd");
+    var condition = new LocalDateTimeEqualCondition(expectedDateTime, "yyyy/MM/dd HH:mm:ss");
 
-    assertThat(condition.toString()).isEqualTo("date value: \"2022/10/11\" (with date value format: \"yyyy/MM/dd\")");
+    assertThat(condition.toString())
+      .isEqualTo("datetime value: \"2022/10/11 12:13:14\" (with datetime value format: \"yyyy/MM/dd HH:mm:ss\")");
   }
 }
