+++ src/test/java/com/codeborne/selenide/conditions/datetime/LocalDateTimeFormatConditionTest.java
@@ -1,11 +1,11 @@
-package com.codeborne.selenide.conditions.date;
+package com.codeborne.selenide.conditions.datetime;
 
 import com.codeborne.selenide.CheckResult;
 import com.codeborne.selenide.Driver;
 import org.junit.jupiter.api.Test;
 import org.openqa.selenium.WebElement;
 
-import java.time.LocalDate;
+import java.time.LocalDateTime;
 
 import static com.codeborne.selenide.CheckResult.Verdict.ACCEPT;
 import static com.codeborne.selenide.CheckResult.Verdict.REJECT;
@@ -15,42 +15,43 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-final class LocalDateFormatConditionTest {
+final class LocalDateTimeFormatConditionTest {
   private final Driver driver = mock();
-  private final LocalDateFormatCondition condition = new LocalDateFormatCondition("yyyy/MM/dd");
+  private final LocalDateTimeFormatCondition condition = new LocalDateTimeFormatCondition("yyyy/MM/dd HH:mm:ss");
   private final WebElement element = mock();
 
   @Test
   void correctValueFormat() {
-    when(element.getAttribute(anyString())).thenReturn("2022/10/11");
+    when(element.getAttribute(anyString())).thenReturn("2022/10/11 12:13:14");
 
     CheckResult check = condition.check(driver, element);
 
     assertThat(check.verdict()).isEqualTo(ACCEPT);
-    assertThat(check.actualValue()).isEqualTo(LocalDate.of(2022, 10, 11));
+    assertThat(check.actualValue()).isEqualTo(LocalDateTime.of(2022, 10, 11, 12, 13, 14));
 
     verify(element).getAttribute("value");
   }
 
   @Test
   void incorrectValueFormat() {
-    when(element.getAttribute(anyString())).thenReturn("2022-10-11");
+    when(element.getAttribute(anyString())).thenReturn("2022-10-11 12:13:14");
 
     CheckResult check = condition.check(driver, element);
 
     assertThat(check.verdict()).isEqualTo(REJECT);
-    assertThat(check.actualValue()).isEqualTo("2022-10-11");
+    assertThat(check.actualValue()).isEqualTo("2022-10-11 12:13:14");
 
     verify(element).getAttribute("value");
   }
 
   @Test
   void formatLocalDateAccordingToHeldPattern() {
-    assertThat(condition.format(LocalDate.of(2022, 10, 11))).isEqualTo("2022/10/11");
+    assertThat(condition.format(LocalDateTime.of(2022, 10, 11, 12, 13, 14)))
+      .isEqualTo("2022/10/11 12:13:14");
   }
 
   @Test
   void stringRepresentationOfCondition() {
-    assertThat(condition.toString()).isEqualTo("date value format: \"yyyy/MM/dd\"");
+    assertThat(condition.toString()).isEqualTo("datetime value format: \"yyyy/MM/dd HH:mm:ss\"");
   }
 }
